requirements.txt
pytest==8.3.2
pytest-xdist==3.6.1
allure-pytest==2.13.5
python-dotenv==1.0.1
pydantic-settings==2.3.4
